feat: Moved and unified container-registry
