+++ src/core/infrastructure/container-registry/logger/logger-module.ts
@@ -1,0 +1,9 @@
+import { Container, ContainerModule } from '../../utils/di/container'
+import { LoggerRepository } from '@/core/domain/repositories/logger/logger-repository'
+import { PinoLoggerRepository } from '@/core/infrastructure/logger/pino-logger-repository'
+import { LoggerAggregator } from '@/core/application/logger/logger-aggregator'
+
+export const LoggerModule = new ContainerModule((container: Container) => {
+  container.bind<LoggerRepository>(LoggerRepository).to(PinoLoggerRepository)
+  container.bind<LoggerAggregator>(LoggerAggregator).toSelf().inSingletonScope()
+})
