+++ src/core/infrastructure/container-registry/midaz-http-fetch-module.ts
@@ -1,5 +1,6 @@
 import { Container, ContainerModule } from '../utils/di/container'
 import { HttpFetchUtils } from '../utils/http-fetch-utils'
+
 export const ContainerTypeMidazHttpFetch = {
   HttpFetchUtils: 'HttpFetchUtils'
 }
