+++ src/core/infrastructure/container-registry/midaz-plugins/auth-module.ts
