+++ src/core/infrastructure/container-registry/midaz-plugins/midaz-plugins-module.ts
@@ -2,7 +2,7 @@ import { Container, ContainerModule } from '../../utils/di/container'
 import { AuthModule } from './auth-module'
 import { IdentityModule } from './identity-module'
 
-export const LerianPluginsModule = new ContainerModule(
+export const MidazPluginsModule = new ContainerModule(
   (container: Container) => {
     container.load(AuthModule)
     container.load(IdentityModule)
