+++ src/core/infrastructure/container-registry/midaz/portfolio-module.ts
@@ -6,11 +6,11 @@ import { UpdatePortfolioRepository } from '@/core/domain/repositories/portfolios
 import { DeletePortfolioRepository } from '@/core/domain/repositories/portfolios/delete-portfolio-repository'
 import { FetchPortfolioByIdRepository } from '@/core/domain/repositories/portfolios/fetch-portfolio-by-id-repository'
 
-import { MidazFetchAllPortfoliosRepository } from '../portfolios/midaz-fetch-all-portfolio-repository'
-import { MidazCreatePortfolioRepository } from '../portfolios/midaz-create-portfolio-repository'
-import { MidazFetchPortfolioByIdRepository } from '../portfolios/midaz-fetch-portfolio-by-id-repository'
-import { MidazUpdatePortfolioRepository } from '../portfolios/midaz-update-portfolio-repository'
-import { MidazDeletePortfolioRepository } from '../portfolios/midaz-delete-portfolio-repository'
+import { MidazFetchAllPortfoliosRepository } from '@/core/infrastructure/midaz/portfolios/midaz-fetch-all-portfolio-repository'
+import { MidazCreatePortfolioRepository } from '@/core/infrastructure/midaz/portfolios/midaz-create-portfolio-repository'
+import { MidazFetchPortfolioByIdRepository } from '@/core/infrastructure/midaz/portfolios/midaz-fetch-portfolio-by-id-repository'
+import { MidazUpdatePortfolioRepository } from '@/core/infrastructure/midaz/portfolios/midaz-update-portfolio-repository'
+import { MidazDeletePortfolioRepository } from '@/core/infrastructure/midaz/portfolios/midaz-delete-portfolio-repository'
 
 export const MidazPortfolioModule = new ContainerModule(
   (container: Container) => {
