+++ src/core/infrastructure/container-registry/midaz/segment-module.ts
@@ -6,11 +6,11 @@ import { FetchAllSegmentsRepository } from '@/core/domain/repositories/segments/
 import { FetchSegmentByIdRepository } from '@/core/domain/repositories/segments/fetch-segment-by-id-repository'
 import { UpdateSegmentRepository } from '@/core/domain/repositories/segments/update-segment-repository'
 
-import { MidazCreateSegmentRepository } from '../segments/midaz-create-segment-repository'
-import { MidazFetchAllSegmentsRepository } from '../segments/midaz-fetch-all-segments-repository'
-import { MidazUpdateSegmentRepository } from '../segments/midaz-update-segment-repository'
-import { MidazDeleteSegmentRepository } from '../segments/midaz-delete-segment-repository'
-import { MidazFetchSegmentByIdRepository } from '../segments/midaz-fetch-segment-by-id-repository'
+import { MidazCreateSegmentRepository } from '@/core/infrastructure/midaz/segments/midaz-create-segment-repository'
+import { MidazFetchAllSegmentsRepository } from '@/core/infrastructure/midaz/segments/midaz-fetch-all-segments-repository'
+import { MidazUpdateSegmentRepository } from '@/core/infrastructure/midaz/segments/midaz-update-segment-repository'
+import { MidazDeleteSegmentRepository } from '@/core/infrastructure/midaz/segments/midaz-delete-segment-repository'
+import { MidazFetchSegmentByIdRepository } from '@/core/infrastructure/midaz/segments/midaz-fetch-segment-by-id-repository'
 
 export const MidazSegmentModule = new ContainerModule(
   (container: Container) => {
