--- src/core/infrastructure/logger/module/logger-module.ts
@@ -1,7 +1,0 @@
-import { LoggerRepository } from '@/core/domain/repositories/logger/logger-repository'
-import { Container, ContainerModule } from '../../utils/di/container'
-import { PinoLoggerRepository } from '../pino-logger-repository'
-
-export const LoggerModule = new ContainerModule((container: Container) => {
-  container.bind<LoggerRepository>(LoggerRepository).to(PinoLoggerRepository)
-})
